requirements.txt
watchdog==3.0.0
kbcstorage==1.3.1
python-dotenv==1.0.0
pyyaml>=6.0.1
python-json-logger>=2.0.7
rumps
pyinstaller
requests>=2.31.0
